+++ lib/constants/splash_string.dart
@@ -1,0 +1,4 @@
+class SplashString {
+  // ignore: constant_identifier_names
+  static const CENTERLOGO = 'VeriFYD.store';
+}