+++ lib/start_app.dart
@@ -1,18 +1,22 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:flutter_svg/svg.dart';
+import 'package:fyd_ui/fyd_ui.dart';
+import 'package:get/get.dart';
 import 'package:verifyd_store/config/app_config.dart';
+import 'package:verifyd_store/utils/router/routes.dart';
 
 class StartApp extends StatelessWidget {
   const StartApp({Key? key}) : super(key: key);
 
-  // This widget is the root of your application.
+  // summation of all widgets heigt must be 640-(24) => 616
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      home: TestScreen(title: AppConfig.of(context).appTitle),
+    return GetMaterialApp(
+      title: 'veriFYD.store',
+      // home: const TestScreen(title: "title"),
+      initialRoute: AppRoutes.getSplashRoute,
+      getPages: AppRoutes.pages,
     );
   }
 }
@@ -23,9 +27,22 @@ class TestScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(title),
+    // all the widgets on this screen must have a height summation of 640-(24) => 636
+    ScreenUtil.init(
+      context,
+      designSize: const Size(360, 640),
+      minTextAdapt: true,
+    );
+
+    return SafeArea(
+      child: Scaffold(
+        backgroundColor: Colors.black,
+        body: Center(
+          child: SvgPicture.asset(
+            'assets/logo/fyd-logo.svg',
+            height: 40.h,
+          ),
+        ),
       ),
     );
   }