+++ lib/utils/router/nav_links.dart
@@ -1,0 +1,6 @@
+class NavLinks {
+  // ignore: constant_identifier_names
+  static const SPLASH = "/splash";
+  // ignore: constant_identifier_names
+  static const HOME = "/home";
+}