+++ CHANGELOG.md
@@ -2,5 +2,14 @@
  * Flavor setup env/prod
  * Firebase setup with sha certificates
  * minSdk set to 21 in android/local.Properties
+
+## 0.1/Splash/UI-F()
+ * branch Added - {Splash/UI-F()}
+ * Packages Added `fyd_ui`, `get`, `animated_text_kit`, `flutter_native_splash`,
+   `flutter_screenutil`, `flutter_svg`
+ * Added Navigation
+ * Added Splash Screen + with Native Splash
+ * Added GetMaterial with Screen util and fyd_ui
+ * Added constants <String> -> Screen wise